admin checks

+++ src/app/components/scheduler/scheduler.component.ts
@@ -45,32 +45,34 @@ import EventUtils from './shared/event.utils';
 
 export class SchedulerComponent implements OnInit {
 
-    @Input() scheduleKey: string;
+    @Input() scheduleKey: string
 
-    loading = true;
-    activeDayIsOpen: boolean = false;
-    view: string = 'month';
-    viewDate: Date = new Date();
+    loading = true
+    isAdmin = false
+    activeDayIsOpen: boolean = false
+    view: string = 'month'
+    viewDate: Date = new Date()
 
-    lunchSchedule: Schedule;
+    calendarEvents: CalendarEvent[] = []
+    events: Event[] = []
+    volunteers: User[] = []
+    lunchSchedule: Schedule
+    userRef: User
 
     actions: CalendarEventAction[] = [
         {
             label: '<i class="material-icons md-18">edit</i>',
             onClick: ({ event }: { event: CalendarEvent }): void => {
-                this.handleEditEvent(event);
+                this.handleEditEvent(event)
             }
         },
         {
             label: '<i class="material-icons md-18">close</i>',
             onClick: ({ event }: { event: CalendarEvent }): void => {
-                this.handleDeleteEvent(event);
+                this.handleDeleteEvent(event)
             }
         }
     ];
-
-    events: CalendarEvent[];
-    volunteers: User[];
 
     constructor(public auth: AuthService,
         private scheduleService: ScheduleService,
@@ -79,43 +81,47 @@ export class SchedulerComponent implements OnInit {
         public snackBar: MatSnackBar) { }
 
     ngOnInit(): void {
-        this.fetchEvents();
-        this.fetchVolunteers();
+        this.fetchEvents()
     }
 
     fetchEvents() {
-        this.scheduleService.getSchedule(this.scheduleKey)
-            .snapshotChanges()
-            .subscribe(data => {
-                var x = data.payload.toJSON();
-                x["$key"] = data.key;
-                this.lunchSchedule = x as Schedule;
-                this.eventService.getScheduleEvents(this.lunchSchedule.$key)
-                    .snapshotChanges()
-                    .subscribe(data => {
-                        this.events = [];
-                        let events = [];
-                        data.forEach(element => {
-                            var x = element.payload.toJSON();
-                            x["$key"] = element.key;
-                            events.push(x as Event);
-                        });
-                        events.forEach((event) => {
-                            this.events.push(EventUtils.mapToCalendarEvent(event));
-                        });
-                        this.events.forEach((event) => {
-                            event["actions"] = this.actions;
-                        });
-                        this.refresh.next();
-                        this.loading = false;
-                    });
-            });
-    }
-
-    fetchVolunteers() {
-        this.auth.getAllVolunteers().subscribe(data => {
-            this.volunteers = data;
-        });
+        this.auth.user$.subscribe(user => {
+            this.userRef = user
+            this.isAdmin = this.auth.isAdmin(user)
+            this.scheduleService.getSchedule(this.scheduleKey)
+                .snapshotChanges()
+                .subscribe(data => {
+                    var x = data.payload.toJSON()
+                    x['$key'] = data.key
+                    this.lunchSchedule = x as Schedule
+                    this.eventService.getScheduleEvents(this.lunchSchedule.$key)
+                        .snapshotChanges()
+                        .subscribe(data => {
+                            data.forEach(element => {
+                                var x = element.payload.toJSON()
+                                x['$key'] = element.key
+                                this.events.push(x as Event)
+                            })
+                            this.events.forEach((event) => {
+                                this.calendarEvents.push(EventUtils.mapToCalendarEvent(event))
+                            })
+                            this.calendarEvents.forEach((event) => {
+                                let target = this.events.find(e => e.$key == event.id)
+                                event['actions'] = (this.isAdmin || target.user == this.userRef.uid) ? this.actions : []
+                            })
+                            this.refresh.next()
+                            this.loading = false
+                            this.auth.getAllVolunteers().subscribe(volunteers => {
+                                if (this.isAdmin) {
+                                    this.volunteers = volunteers
+                                } else {
+                                    this.volunteers = []
+                                    this.volunteers.push(volunteers.find(e => e.uid == this.userRef.uid))
+                                }
+                            })
+                        })
+                })
+        })
     }
 
     dayClicked({ date, events }: { date: Date; events: Array<CalendarEvent<{ film: Event }>>; }): void {
@@ -124,15 +130,15 @@ export class SchedulerComponent implements OnInit {
                 (isSameDay(this.viewDate, date) && this.activeDayIsOpen === true) ||
                 events.length === 0
             ) {
-                this.activeDayIsOpen = false;
+                this.activeDayIsOpen = false
             } else {
-                this.activeDayIsOpen = true;
-                this.viewDate = date;
+                this.activeDayIsOpen = true
+                this.viewDate = date
             }
         }
     }
 
-    refresh: Subject<any> = new Subject();
+    refresh: Subject<any> = new Subject()
 
     eventTimesChanged({ event, newStart, newEnd }: CalendarEventTimesChangedEvent, user): void {
         event.start = newStart;
@@ -150,39 +156,45 @@ export class SchedulerComponent implements OnInit {
     }
 
     handleEditEvent(event: CalendarEvent): void {
-        let dialogRef = this.dialog.open(EventDialog, {
-            width: '650px',
-            data: { event: event, volunteers: this.volunteers }
-        });
+        let target = this.events.find(e => e.$key == event.id)
+        if (this.isAdmin || target.user == this.userRef.uid) {
+            let dialogRef = this.dialog.open(EventDialog, {
+                width: '650px',
+                data: { event: event, volunteers: this.volunteers }
+            });
 
-        dialogRef.afterClosed().subscribe(result => {
-            if (result) {
-                delete result["$key"];
-                result["schedule_key"] = this.scheduleKey;
-                this.eventService.updateEvent(event.id, result)
-                    .then((data) => {
-                        this.openSnackBar('Schedule Saved', 'OKAY');
-                    })
-                    .catch((error) => {
-                        this.openSnackBar(error, 'OKAY');
-                    });
-            }
-        });
+            dialogRef.afterClosed().subscribe(result => {
+                if (result) {
+                    delete result["$key"];
+                    result["schedule_key"] = this.scheduleKey;
+                    this.eventService.updateEvent(event.id, result)
+                        .then((data) => {
+                            this.openSnackBar('Schedule Saved', 'OKAY');
+                        })
+                        .catch((error) => {
+                            this.openSnackBar(error, 'OKAY')
+                        })
+                }
+            })
+        }
     }
 
     handleDeleteEvent(event: CalendarEvent): void {
-        let dialogRef = this.dialog.open(EventDeleteDialog, {
-            width: '450px',
-            data: { event: event }
-        });
+        let target = this.events.find(e => e.$key == event.id)
+        if (this.isAdmin || target.user == this.userRef.uid) {
+            let dialogRef = this.dialog.open(EventDeleteDialog, {
+                width: '450px',
+                data: { event: event }
+            });
 
-        dialogRef.afterClosed().subscribe(result => {
-            if (result) {
-                this.eventService.deleteEvent(event.id as string);
-                this.events = this.events.filter(iEvent => iEvent !== event);
-                this.openSnackBar('Schedule Removed', 'OKAY');
-            }
-        });
+            dialogRef.afterClosed().subscribe(result => {
+                if (result) {
+                    this.eventService.deleteEvent(event.id as string);
+                    this.calendarEvents = this.calendarEvents.filter(iEvent => iEvent !== event);
+                    this.openSnackBar('Schedule Removed', 'OKAY');
+                }
+            })
+        }
     }
 
     handleCreateEvent() {
